Revert "Remove unnecessary conditional check for temporary directory cleanup"
This reverts commit 30f8c6c64fea44e2c36be67ddbd26eb294179f54.

+++ internal/application/handlers/git_http.go
@@ -111,7 +111,9 @@ func GitHTTPBackend(c *gin.Context) {
 			renamePlatformHubFiles(tempRepoPath, gitPath, processTemplatesDir, username, password)
 			renamePlatformHubFiles(tempRepoPath, gitPath, policiesDir, username, password)
 		}
+	}
 
+	if created {
 		defer func() {
 			err := os.RemoveAll(tempRepoPath)
 			if err != nil {
